creating parsing functions

+++ src/defaults.h
@@ -69,7 +69,8 @@ public:
 	static const int COTOP = 11;
 	static const int LOGOP = 12;
 	static const int NOPERS = 13;
-	static const int m_size = 13;
+	
+	static const size_t m_size = 13;
 	
 	static operation <operand <data_t>> opers[];
 };
